requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
mem0ai
langchain==0.1.0
openai==1.6.1
anthropic==0.7.7
PyGithub==2.1.1
PyPDF2==3.0.1
python-dotenv==1.0.0
Authlib==1.3.0
requests==2.31.0
langchain-openai
